refactor: refactor product page info, more code splittling
