+++ src/components/PageMain/components/MainFilters/components/SlideFilter/MainSlideFilter.ts
@@ -1,8 +1,8 @@
 import { createElem } from '../../../../../../utils/create-element';
 import { createInput } from '../../../../../../utils/create-input-element';
 import styles from './MainSlideFilter.module.scss';
-import { appliedFilters } from '../../../../../../const/store';
-import { renderFiltered } from '../../../Filter/filter';
+import { onChangeRangeInput } from './handlers/onChangeRangeInput';
+import { onChangeRangeNumInput } from './handlers/onChangeRangeNumInput';
 
 /** Функция создает универсальный фильтр со слайдером, аргументы: название фильтра, подкласс и массив чисел для фильтрации */
 export const renderSlideFilter = (title: string, rangeIcon: string, data: number[], subClass: string): HTMLElement => {
@@ -42,42 +42,7 @@ export const renderSlideFilter = (title: string, rangeIcon: string, data: number
 
   slideFilterNumbers.append(numbersFrom, numbersTo);
 
-  const priceGap = 1;
-
-  slideFilterNumbers.onchange = (e: Event): void => {
-    const targetInput = e.target as HTMLInputElement;
-    const filterType =
-      targetInput.classList.contains('num-input-left_price') || targetInput.classList.contains('num-input-right_price')
-        ? 'price'
-        : 'stock';
-
-    const side = targetInput.className.includes('left') ? 'left' : 'right';
-
-    const rangeInputsParent = document.querySelector(`.slide-filter__ranges_${filterType}`) as HTMLElement;
-    const progress = document.querySelector(`.progress_${filterType}`) as HTMLElement;
-    const rangeInputs = rangeInputsParent.children;
-
-    const rangeLeft = rangeInputs[0] as HTMLInputElement;
-    const rangeRight = rangeInputs[1] as HTMLInputElement;
-
-    const minVal = side === 'left' ? +targetInput.value : +rangeLeft.value;
-    const maxVal = side === 'right' ? +targetInput.value : +rangeRight.value;
-
-    if (maxVal - minVal < priceGap) {
-      if (side === 'left') {
-        rangeLeft.value = (maxVal - priceGap).toString();
-      } else {
-        rangeRight.value = (minVal + priceGap).toString();
-      }
-    } else {
-      progress.style.left = `${(minVal / +rangeLeft.max) * 100}%`;
-      progress.style.right = `${100 - (maxVal / +rangeRight.max) * 100}%`;
-    }
-
-    if (!appliedFilters[filterType]) appliedFilters[filterType] = [];
-    appliedFilters[filterType] = [minVal, maxVal];
-    renderFiltered(appliedFilters);
-  };
+  slideFilterNumbers.onchange = onChangeRangeNumInput;
 
   // Double range slider
   const slideFilterSlider: HTMLElement = createElem('div', 'slide-filter__slider');
@@ -102,35 +67,7 @@ export const renderSlideFilter = (title: string, rangeIcon: string, data: number
   rightRangeInput.setAttribute('max', max);
   rightRangeInput.setAttribute('value', max);
 
-  rangeInputs.oninput = (e: Event): void => {
-    const targetInput = e.target as HTMLInputElement;
-    const rangeInputsEl = targetInput.parentElement as HTMLElement;
-    const filterType = rangeInputsEl.classList.contains('slide-filter__ranges_price') ? 'price' : 'stock';
-    const beforRangeInputParent = rangeInputsEl.previousElementSibling as HTMLElement;
-    const progress = beforRangeInputParent.firstElementChild as HTMLElement;
-
-    const rangeInputs = rangeInputsEl.children;
-    const rangeLeft = rangeInputs[0] as HTMLInputElement;
-    const rangeRight = rangeInputs[1] as HTMLInputElement;
-
-    const minVal = +rangeLeft.value;
-    const maxVal = +rangeRight.value;
-
-    if (maxVal - minVal < priceGap) {
-      if (targetInput.classList.contains('range-input_left')) {
-        rangeLeft.value = (maxVal - priceGap).toString();
-      } else {
-        rangeRight.value = (minVal + priceGap).toString();
-      }
-    } else {
-      progress.style.left = `${(minVal / +rangeLeft.max) * 100}%`;
-      progress.style.right = `${100 - (maxVal / +rangeRight.max) * 100}%`;
-    }
-
-    if (!appliedFilters[filterType]) appliedFilters[filterType] = [];
-    appliedFilters[filterType] = [minVal, maxVal];
-    renderFiltered(appliedFilters);
-  };
+  rangeInputs.oninput = onChangeRangeInput;
 
   rangeInputs.append(leftRangeInput, rightRangeInput);
   slideFilter.append(heading, slideFilterNumbers, slideFilterSlider, rangeInputs);
