+++ src/components/PageProducts/conponents/Product/components/Information/components/InfoAboutProducts/InfoAboutProducts.module.scss
@@ -1,22 +1,5 @@
 @import '/src/theme/variable';
 
-.product-page__header-wrapper {
-  margin-bottom: 2rem;
-}
-
-.product-page__header {
-  display: block;
-  text-align: center;
-  font-size: 24px;
-}
-
-.product-page__header-descrip {
-  display: block;
-  text-align: center;
-  font-size: 14px;
-  color: $light-text;
-}
-
 .product-page__data {
   position: relative;
   margin-top: 2rem;
@@ -34,10 +17,6 @@
   line-height: 1.7rem;
   font-weight: 400;
 }
-.item-descriprions {
-  display: block;
-  position: relative;
-}
 
 .product-page__data-item-title,
 .product-page__data-item-value {
@@ -47,6 +26,11 @@
 
 .product-page__data-item-title {
   color: $light-text;
+  position: relative;
+}
+
+.item-descriprions {
+  display: block;
   position: relative;
 }
 
@@ -125,14 +109,4 @@
 .show_content {
   overflow: hidden;
   height: 332px;
-}
-
-.product-page__price {
-  padding-top: 1rem;
-  margin-bottom: 1rem;
-  font-size: 1.4rem;
-  line-height: 2.3rem;
-  font-weight: 400;
-  text-align: center;
-  border-top: 1px solid $primaty-5;
 }
