+++ src/components/PageProducts/conponents/Product/components/Information/components/InfoPriceProduct/InfoPriceProduct.module.scss
@@ -1,0 +1,11 @@
+@import '/src/theme/variable';
+
+.product-page__price {
+  padding-top: 1rem;
+  margin-bottom: 1rem;
+  font-size: 1.4rem;
+  line-height: 2.3rem;
+  font-weight: 400;
+  text-align: center;
+  border-top: 1px solid $primaty-5;
+}
