+++ src/components/PageProducts/conponents/Product/components/Information/components/InfoTitle/InfoTitle.module.scss
@@ -1,0 +1,18 @@
+@import '/src/theme/variable';
+
+.product-page__header-wrapper {
+  margin-bottom: 2rem;
+}
+
+.product-page__header {
+  display: block;
+  text-align: center;
+  font-size: 24px;
+}
+
+.product-page__header-descrip {
+  display: block;
+  text-align: center;
+  font-size: 14px;
+  color: $light-text;
+}
